Fix paging + user mode + syscall integration
- Fix user program EIP/ESP virtual address mapping
- Add VGA buffer mapping to user address space
- Fix kernel memory permissions for user program static variables
- Enhance page fault debugging with detailed information
- Improve scheduler stability and console compatibility

Enables working user mode processes with system calls under paging.

+++ include/arch/x86/idt.hpp
@@ -117,7 +117,7 @@ public:
      * @param interruptNumber Interrupt number
      * @return Pointer to IDT entry or nullptr if invalid
      */
-    static const IDTEntry* get_entry(u8 interruptNumber);
+    static const IDTEntry* get_entry(u32 interruptNumber);
 };
 
 } // namespace kira::system
